api/python/requirements.txt
python-dotenv==1.0.0
python-http-client>=3.3.5
http-server>=0.12.0 